qcacld-3.0: Add debugging infrastructure to Android.mk
The qcacld-3.0 Android.mk is unique in that it can be called
recursively. This behavior can make it difficult to debug build
issues. To help alleviate this problem, add a user function to
Android.mk to allow debug information to be logged, and add a few logs
which will be useful in debugging recursion issues. In addition change
a few $(warning) calls to use the infrastructure since under normal
use the Android.mk should be silent.

Change-Id: I660ceddaf53e42c331deba5d33b32360b549cfc7
CRs-Fixed: 2890484
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>