msm: camera: isp: Add isp resource name
For ISP resources, res_id is printed, which does not give clear
indication of resource. At some places translation code is
used which is an overhead.
This commit adds name to isp resource node structure.
This will be helpful while printing the resource logs and
avoids the translation overhead.

CRs-Fixed: 2830502
Change-Id: I436810e971a521fe2e845a2e3ca544b4c2921727
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>