qcacld-3.0: Use static symbolic links for multi-ko support
The current multi-ko support is achieved by creating, during the
actual build process, target-specific .<chip> symbolic links beneath
the qcacld-3.0 folder. Unfortunately this mechanism violates the
requirements of the Hermetic Build feature which requires that the
source directory not be modified during the build process.

To satisfy this requirement remove the dynamic creation of symbolic
links from Android.mk, and instead staticially instantiate all of the
required symbolic links in the git project.

Change-Id: I7d75bb6a82fd27d8925d53529b4252b782457bc8
CRs-Fixed: 2896976
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>