qcacmn: Add sanity check for the command to remove in connection mgr
Add sanity check for the command to remove in connection mgr, as
by default in case of invalid cm id the disconnetc command is removed.

Also cm_id pointer passed in the cm_remove_cmd, may get freed as
part of cm_delete_req_from_list(), if its part of the cm_req.

Also move the add to ser logic for the roam sync to connect state
handling.

Change-Id: Ie69e8a65fc864a55c51d149570573e6f9aef4564
CRs-Fixed: 2911267
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>