qcacld-3.0: Enabled nested composites on supported GKI kernels
KERNEL_SUPPORTS_NESTED_COMPOSITES := y is used to enable nested
composite support. The nested composite support is available in some
MSM kernels, and is available in all GKI kernels beginning with
5.10.20, but unfortunately is not available in any upstream kernel.

When the feature is present in an MSM kernel, the flag is explicitly
set in the kernel sources.  When a GKI kernel is used, there isn't a
flag set in the sources. Therefore add logic to set the flag if we are
building with GKI kernel 5.10.20 or greater.

Change-Id: I13cd1a07915e9d187da22259cc61d9481e908e79
CRs-Fixed: 2914541
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>