msm: camera: isp: Mask process_bubble for SHDR usecase
SHDR usecase doesn't report bubble to CRM, but we still
need to know whether it is processing bubble during
applying request, since we may haven't received the
reg update of the previous applied request, then
we have a chance to apply a new req to HW, this may
result in overflow issue. This change masks the
process_bubble for SHDR case during epoch in applied
sub state.

CRs-Fixed: 2948932
Change-Id: I699a2f59ea874693b78a918895db98727b95c634
Signed-off-by: Depeng Shao <depengs@codeaurora.org>